+++ src/main/java/com/netgrif/application/engine/petrinet/service/ProcessRoleService.java
@@ -240,6 +240,11 @@ public class ProcessRoleService implements IProcessRoleService {
     }
 
     @Override
+    public Set<ProcessRole> findAllGlobalRoles() {
+        return processRoleRepository.findAllByGlobalIsTrue();
+    }
+
+    @Override
     public List<ProcessRole> findAll(String netId) {
         Optional<PetriNet> netOptional = netRepository.findById(netId);
         if (netOptional.isEmpty())