+++ src/main/java/com/netgrif/application/engine/petrinet/service/interfaces/IProcessRoleService.java
@@ -28,6 +28,8 @@ public interface IProcessRoleService {
 
     List<ProcessRole> findAll();
 
+    Set<ProcessRole> findAllGlobalRoles();
+
     List<ProcessRole> findAll(String netId);
 
     ProcessRole defaultRole();