+++ src/main/java/com/netgrif/application/engine/workflow/service/interfaces/IDataService.java
@@ -99,7 +99,7 @@ public interface IDataService {
 
     void validateCaseRefValue(List<String> value, List<String> allowedNets) throws IllegalArgumentException;
 
-    void validateTaskRefValue(List<String> value, String restrictedTaskId) throws IllegalArgumentException;
+//    void validateTaskRefValue(List<String> value, String restrictedTaskId) throws IllegalArgumentException;
 
     SetDataEventOutcome changeComponentProperties(Case useCase, String transitionId, String fieldId, Map<String, String> properties);